Add User register view and url conf, checks authorization for dogs

+++ bloodline_backend/bloodline_api/views/dog.py
@@ -4,7 +4,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status, permissions
 from bloodline_api.models import Dog
-from bloodline_api.serializers import DogSerializer
+from bloodline_api.serializers import DogSerializer, DogCreateSerializer, DogPutSerializer
 from bloodline_api.permissions import IsOwnerOrReadOnly
 
 #
@@ -16,17 +16,18 @@ class DogList(APIView):
   """
   permission_classes = [permissions.IsAuthenticatedOrReadOnly]
     
-  def get(self, request, format=None):
+  def get(self, request):
     dogs = Dog.objects.all()
     serializer = DogSerializer(dogs, many=True)
     return Response(serializer.data)
 
-  def post(self, request, format=None):
-    serializer = DogSerializer(data=request.data)
+  def post(self, request):
+    serializer = DogCreateSerializer(data=request.data)
     if serializer.is_valid():
-      serializer.save()
+      dog = serializer.save(owner=request.user)
       return Response(serializer.data, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    else:
+      return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 class DogDetail(APIView):
   """
@@ -50,7 +51,7 @@ class DogDetail(APIView):
   def put(self, request, pk, format=None):
     dog = self.get_object(pk)
     self.check_object_permissions(request, dog)
-    serializer = DogSerializer(dog, data=request.data)
+    serializer = DogPutSerializer(dog, data=request.data)
     if serializer.is_valid():
       serializer.save()
       return Response(serializer.data)
@@ -58,6 +59,7 @@ class DogDetail(APIView):
 
   def delete(self, request, pk, format=None):
     dog = self.get_object(pk)
+    self.check_object_permissions(request, dog)
     dog.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
